backend/requirements.txt
# FastAPI and Server Dependencies
fastapi==0.115.5
uvicorn[standard]==0.32.0
starlette==0.41.2
pydantic==2.9.2
python-multipart==0.0.17

# Google AI Dependencies
google-generativeai==0.8.3
google-api-core==2.23.0
google-auth==2.36.0
google-api-python-client==2.152.0

# SSH and File Transfer
paramiko==3.5.0
scp==0.15.0
bcrypt==4.2.0
cryptography==43.0.3
PyNaCl==1.5.0

# HTTP and Networking
requests==2.32.3
urllib3==2.2.3
httplib2==0.22.0

# Utility Dependencies
typing_extensions==4.12.2
anyio==4.6.2.post1
click==8.1.7
tqdm==4.67.0

# Data Processing
protobuf==5.28.3
grpcio==1.67.1
pydantic_core==2.23.4